+++ packages/widgetbook_generator/lib/src/core/story_class_builder.dart
@@ -55,6 +55,16 @@ class StoryClassBuilder {
       isNullable: true,
     );
 
+    final componentClassRef = TypeReference(
+      (b) =>
+          b
+            ..symbol = 'Component'
+            ..types.addAll([
+              widgetClassRef,
+              argsClassRef,
+            ]),
+    );
+
     return Class(
       (b) =>
           b
@@ -190,7 +200,22 @@ class StoryClassBuilder {
                 },
               ),
             )
-            ..methods.add(
+            ..methods.addAll([
+              Method(
+                (b) =>
+                    b
+                      ..name = 'component'
+                      ..type = MethodType.getter
+                      ..returns = componentClassRef
+                      // Casting is needed for generic stories.
+                      // A Story<Widget<T>, Args<T>> has a Component<Widget<T>, Args<T>>
+                      // but the generated component variable is of type
+                      // Component<Widget<dynamic>, Args<dynamic>>
+                      ..body =
+                          refer(
+                            '${widgetType.nonGenericName}Component',
+                          ).asA(componentClassRef).code,
+              ),
               Method(
                 (b) =>
                     b
@@ -223,7 +248,7 @@ class StoryClassBuilder {
                             },
                           ).code,
               ),
-            ),
+            ]),
     );
   }